+++ src/chained_powell.jl
@@ -16,7 +16,7 @@ function chained_powell_model(N = 1000; T = Float64; backend = CUDABackend())
     )
 
     ExaModels.objective(c, (x[2*i-1] + 10 * x[2*i])^2 + 5 * (x[2*i+1] - x[2*i+2])^2 + 
-    (x[2*i] - 2 * x[2*i+1])^4 + 10 * (x[2*i-1] - x[2*i+2])^4 for i = 1:N/2-1)
+    (x[2*i] - 2 * x[2*i+1])^4 + 10 * (x[2*i-1] - x[2*i+2])^4 for i = 1:Int(N/2-1))
     return ExaModels.ExaModel(c)
 end
